refactor: Parser 클래스명 수정
